requirements.txt
asgiref==3.5.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.12
django-cors-headers==3.11.0
django-import-export==2.7.1
djongo==1.3.6
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
psycopg2==2.9.3
pymongo==3.12.3
pytz==2021.3
PyYAML==6.0
sqlparse==0.2.4
tablib==3.2.0
xlrd==2.0.1
xlwt==1.3.0
